+++ rwengine/src/objects/PickupObject.cpp
@@ -191,9 +191,9 @@ void PickupObject::tick(float dt) {
     float time = engine->getGameTime();
     float colourValue = 0.5f * (std::sin(time * 3.0664064f) * 0.3f + 0.3f);
     uint32_t* colour = &colours[m_colourId];
-    float red = (*colour >> 16) & 0xFF;
-    float green = (*colour >> 8) & 0xFF;
-    float blue = *colour & 0xFF;
+    float red = static_cast<float>((*colour >> 16) & 0xFF);
+    float green = static_cast<float>((*colour >> 8) & 0xFF);
+    float blue = static_cast<float>(*colour & 0xFF);
     m_corona.colour =
         glm::vec4(red / 255.f, green / 255.f, blue / 255.f, 1.f) * colourValue;
